external/japanese_lexical_network/requirements.txt
click
colorama
Flask
itsdangerous
Jinja2
MarkupSafe
networkx
setuptools==75.1.0
Werkzeug
wheel==0.44.0
SPARQLWrapper
requests
google-generativeai
pillow
python-dotenv
redis
flask_caching
jreadability
gunicorn
pandas
tabula-py
PyMuPDF
networkx
matplotlib